Moved the new classes to some more appropriate packages
