+++ src/main/java/org/eln2/mc/Eln2.kt
@@ -9,7 +9,7 @@ import org.eln2.mc.client.events.ClientEvents
 import org.eln2.mc.common.Configuration
 import org.eln2.mc.common.ElectricalAgeConfiguration
 import org.eln2.mc.common.blocks.BlockRegistry
-import org.eln2.mc.common.blocks.PartRegistry
+import org.eln2.mc.common.parts.PartRegistry
 import org.eln2.mc.common.cell.CellRegistry
 import org.eln2.mc.common.containers.ContainerRegistry
 import org.eln2.mc.common.items.ItemRegistry
