+++ src/main/java/org/eln2/mc/common/blocks/MultipartBlockEntity.kt
@@ -4,6 +4,7 @@ import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
 import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.state.BlockState
+import org.eln2.mc.common.parts.Part
 
 /**
  * Multipart entities
